requirements.txt
flask
waitress
python-dotenv
auto-lazy-imports
pandas